531-django-series/531m.dedal/requirements.txt

django>=2,<2.2

python-dotenv==0.10.3

django-admin-generator

django-dedal

whitenoise==4.1.2
dj-database-url==0.5.0
django-bootstrap-form==3.4

#psycopg2-binary==2.7.6.1